requirements.txt
asgiref==3.11.1
Django==6.0.3
django-axes==8.3.1
django-cors-headers==4.9.0
django-extensions==4.1
django-ipware==7.0.1
django-user-agents==0.4.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.8
PyJWT==2.12.1
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3
PyMySQL==1.1.2
gunicorn==23.0.0
Pillow==11.2.1
whitenoise==6.9.0
django-redis==5.4.0
